# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Me Socorre API",
    "description": "Documentação da API da plataforma de prestação de serviços Me Socorre.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Ambiente local"
    }
  ],
...
  "paths": {
...
    "/admin/providers": {
      "get": {
        "tags": [
          "Administração"
        ],
        "summary": "Lista os prestadores para administração",
        "description": "Lista todos os prestadores, com filtro opcional pelo status de aprovação.",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "enum": {
                  "type": "array",
                  "example": [
                    "PENDING",
                    "APPROVED",
                    "REJECTED",
                    "SUSPENDED",
                    "BANNED"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "description": "Status de aprovação do prestador."
          }
        ],
        "responses": {
          "200": {
            "description": "Prestadores retornados com sucesso."
          },
          "400": {
            "description": "Filtro de status inválido."
          },
          "401": {
            "description": "Token não informado, inválido ou expirado."
          },
          "403": {
            "description": "Acesso permitido apenas para administradores."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/providers/": {
      "get": {
        "tags": [
          "Prestadores"
        ],
        "summary": "Lista os prestadores disponíveis",
        "description": "Retorna prestadores aprovados e disponíveis, com filtros opcionais.",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Busca por nome, categoria ou endereço."
          },
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "minimum": {
                  "type": "number",
                  "example": 1
                }
              }
            },
            "description": "Filtra pelo ID da categoria."
          }
        ],
        "responses": {
          "200": {
            "description": "Prestadores retornados com sucesso."
          }
        }
      }
    },
    "/providers/{id}/reviews": {
      "get": {
        "tags": [
          "Prestadores"
        ],
        "summary": "Lista as avaliações de um prestador",
        "description": "Retorna a média e as avaliações de um prestador aprovado.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "format": {
                  "type": "string",
                  "example": "uuid"
                }
              }
            },
            "description": "ID do perfil do prestador."
          }
        ],
        "responses": {
          "200": {
            "description": "Avaliações retornadas com sucesso."
          },
          "400": {
            "description": "ID inválido."
          },
          "404": {
            "description": "Prestador não encontrado."
          }
        }
      }
    },
    "/providers/{id}/review-summary": {
      "get": {
        "tags": [
          "Prestadores"
        ],
        "summary": "Gera o resumo das avaliações",
        "description": "Gera com IA um resumo das avaliações de um prestador aprovado.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "format": {
                  "type": "string",
                  "example": "uuid"
                }
              }
            },
            "description": "ID do perfil do prestador."
          }
        ],
        "responses": {
          "200": {
            "description": "Resumo retornado com sucesso."
          },
          "400": {
            "description": "ID inválido."
          },
          "404": {
            "description": "Prestador não encontrado."
          },
          "503": {
            "description": "Serviço de geração do resumo indisponível."
          }
        }
      }
    },
    "/providers/{id}": {
      "get": {
        "tags": [
          "Prestadores"
        ],
        "summary": "Consulta um prestador",
        "description": "Retorna o perfil público de um prestador aprovado.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "format": {
                  "type": "string",
                  "example": "uuid"
                }
              }
            },
            "description": "ID do perfil do prestador."
          }
        ],
        "responses": {
          "200": {
            "description": "Prestador encontrado."
          },
          "404": {
            "description": "Prestador não encontrado."
          }
        }
      },
      "put": {
        "tags": [
          "Prestadores"
        ],
        "summary": "Atualiza um prestador",
        "description": "Atualiza o próprio perfil do prestador. Administradores podem atualizar qualquer prestador.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "format": {
                  "type": "string",
                  "example": "uuid"
                }
              }
            },
            "description": "ID do perfil do prestador."
          }
        ],
        "responses": {
          "200": {
            "description": "Prestador atualizado com sucesso."
          },
          "400": {
            "description": "ID ou dados inválidos."
          },
          "401": {
            "description": "Token não informado, inválido ou expirado."
          },
          "403": {
            "description": "Usuário sem permissão para atualizar o prestador."
          },
          "404": {
            "description": "Prestador ou categoria não encontrado."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string",
                    "minLength": 2,
                    "example": "Maria Souza"
                  },
                  "telefone": {
                    "type": "string",
                    "nullable": true,
                    "minLength": 8,
                    "maxLength": 30,
                    "example": "51988888888"
                  },
                  "endereco": {
                    "type": "string",
                    "nullable": true,
                    "minLength": 5,
                    "example": "Av. Central, 250"
                  },
                  "idCategoria": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                  }
                }
              }
            }
          }
        }
      }
    },
    "/providers/{id}/status": {
      "patch": {
        "tags": [
          "Prestadores"
        ],
        "summary": "Altera a disponibilidade do prestador",
        "description": "Ativa ou desativa a disponibilidade de um prestador aprovado.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "format": {
                  "type": "string",
                  "example": "uuid"
                }
              }
            },
            "description": "ID do perfil do prestador."
          }
        ],
        "responses": {
          "200": {
            "description": "Disponibilidade atualizada com sucesso."
          },
          "400": {
            "description": "ID ou dados inválidos."
          },
          "401": {
            "description": "Token não informado, inválido ou expirado."
          },
          "403": {
            "description": "Usuário sem permissão para alterar a disponibilidade."
          },
          "404": {
            "description": "Prestador não encontrado."
          },
          "409": {
            "description": "Prestador ainda não foi aprovado."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "disponivel"
                ],
                "properties": {
                  "disponivel": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/requests/{id}/provider": {
      "patch": {
        "tags": [
          "Solicitações"
        ],
        "summary": "Aceita uma solicitação",
        "description": "Permite que o prestador vinculado aceite uma solicitação pendente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "minimum": {
                  "type": "number",
                  "example": 1
                }
              }
            },
            "description": "ID da solicitação."
          }
        ],
        "responses": {
          "200": {
            "description": "Solicitação aceita com sucesso."
          },
          "400": {
            "description": "ID inválido."
          },
          "401": {
            "description": "Token não informado, inválido ou expirado."
          },
          "403": {
            "description": "Prestador não aprovado ou indisponível."
          },
          "409": {
            "description": "Solicitação indisponível para aceite."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Informe o token JWT retornado pelo endpoint de login."
      }
    }
  }
}